+++ src/app/shared/easter-egg/konamiCode.ts
