+++ src/index.tsx
@@ -1,4 +1,4 @@
-import './app/shared/easterEgg/konamiCode';
+import './app/shared/easter-egg/konamiCode';
 import App from './app/App';
 import AuthenticationProvider from './app/components/provider/AuthenticationProvider';
 import ContextProvider from './app/components/provider/ContextProvider';
